refactor(data): shrink event-reader/status-store functions and split shared_repo module
