+++ src/quodeq/data/events/reader.py
@@ -13,6 +13,57 @@ from quodeq.data.events.codec import EventDecodeError, event_from_dict
 _logger = logging.getLogger(__name__)
 
 
+def _parse_event_line(
+    line: str, log_path: Path, line_num: int, since_timestamp: Optional[datetime],
+) -> Optional[BaseEvent]:
+    """Parse+decode one JSONL line into an event, or None if it should be
+    skipped (a warning/error is already logged for every skip)."""
+    try:
+        # 1. Parse as raw dict first to determine the type
+        raw_data = json.loads(line)
+        event_type_str = raw_data.get("event_type")
+
+        if not event_type_str:
+            _logger.warning(f"Missing 'event_type' in {log_path} at line {line_num}")
+            return None
+
+        # 2. Use the map to find the correct event model
+        # We must convert the string to the Enum member
+        from quodeq.core.events.models import EventType
+        event_type = EventType(event_type_str)
+
+        model_cls = EVENT_MODEL_MAP.get(event_type)
+        if not model_cls:
+            _logger.warning(f"No model mapped for event type {event_type_str} in {log_path} at line {line_num}")
+            return None
+
+        # 3. Decode the full event against the specific model
+        event = event_from_dict(model_cls, raw_data)
+
+        # 4. Checkpoint logic (using strict inequality)
+        if since_timestamp and event.timestamp <= since_timestamp:
+            return None
+
+        return event
+
+    except json.JSONDecodeError:
+        # Must precede (ValueError, KeyError): JSONDecodeError is a
+        # ValueError subclass and would otherwise never match here.
+        _logger.error(f"Malformed JSON in {log_path} at line {line_num}")
+        return None
+    except EventDecodeError as e:
+        # Also a ValueError subclass, so it must precede that catch.
+        _logger.error(f"Schema mismatch in {log_path} at line {line_num}: {e}")
+        return None
+    except (ValueError, KeyError) as e:
+        # Handles invalid Enum values or missing keys in raw_data
+        _logger.error(f"Invalid event structure in {log_path} at line {line_num}: {e}")
+        return None
+    except Exception as e:
+        _logger.error(f"Unexpected error reading {log_path} at line {line_num}: {e}")
+        return None
+
+
 class EventLogReader:
     """
     A streaming reader for the Quodeq Event Log (JSONL).
@@ -44,51 +95,9 @@ class EventLogReader:
                 line = line.strip()
                 if not line:
                     continue
-
-                try:
-                    # 1. Parse as raw dict first to determine the type
-                    raw_data = json.loads(line)
-                    event_type_str = raw_data.get("event_type")
-                    
-                    if not event_type_str:
-                        _logger.warning(f"Missing 'event_type' in {self.log_path} at line {line_num}")
-                        continue
-                    
-                    # 2. Use the map to find the correct event model
-                    # We must convert the string to the Enum member
-                    from quodeq.core.events.models import EventType
-                    event_type = EventType(event_type_str)
-
-                    model_cls = EVENT_MODEL_MAP.get(event_type)
-                    if not model_cls:
-                        _logger.warning(f"No model mapped for event type {event_type_str} in {self.log_path} at line {line_num}")
-                        continue
-
-                    # 3. Decode the full event against the specific model
-                    event = event_from_dict(model_cls, raw_data)
-                    
-                    # 4. Checkpoint logic (using strict inequality)
-                    if since_timestamp and event.timestamp <= since_timestamp:
-                        continue
-
+                event = _parse_event_line(line, self.log_path, line_num, since_timestamp)
+                if event is not None:
                     yield event
-
-                except json.JSONDecodeError:
-                    # Must precede (ValueError, KeyError): JSONDecodeError is a
-                    # ValueError subclass and would otherwise never match here.
-                    _logger.error(f"Malformed JSON in {self.log_path} at line {line_num}")
-                    continue
-                except EventDecodeError as e:
-                    # Also a ValueError subclass, so it must precede that catch.
-                    _logger.error(f"Schema mismatch in {self.log_path} at line {line_num}: {e}")
-                    continue
-                except (ValueError, KeyError) as e:
-                    # Handles invalid Enum values or missing keys in raw_data
-                    _logger.error(f"Invalid event structure in {self.log_path} at line {line_num}: {e}")
-                    continue
-                except Exception as e:
-                    _logger.error(f"Unexpected error reading {self.log_path} at line {line_num}: {e}")
-                    continue
 
     def read_all(self, since_timestamp: Optional[datetime] = None) -> List[BaseEvent]:
         """Convenience method to read all available events into a list."""
